Refactored CGD domain out of this module
The domain classes were put in the cgd module, on the other hand the
classe that uses the SyncRegistrationWithExternalServices interface had
to be moved into this module, since the interface is specified in here.

This commit also contains a small migration code in
FenixeduUlisboaSpecificationsInitializer which will then be removed
after roll out.