Changes StudentActive algorithm for 1stYear1stTime
The comment in the code already explains the issue. We have some DGES
students that cancel the registration but StudentActive is still
reporting them as active because they have a studentCandidacy.

We can also not check only if the registration is active since the DGES
importer, imports the candicies initially with a inactive state, so we
have to check for that specific state as well.

Relates with: #UL-FF-3273