Adds class with @BennuSpringModule for spaces
Spring controllers are only registered in bennu when there's a @BennuSpringModule
that identifies the package of the controller in the basePackages.

Since we are removing bennu-spring from spaces, I'm moving the class
into specifications. Creating a new class instead of having multiple
packages defined in FenixeduUlisboaSpecificationsSpringConfiguration so
we can see in the UI two different applications.

Relates with: #qubIT-Omnis-5811